step4 refactor: directory명 변경
