+++ src/main/kotlin/lotto/vo/Lotto.kt
@@ -1,4 +1,4 @@
-package autolotto.vo
+package lotto.vo
 
 class Lotto(originNumbers: List<LottoNumber> = listOf()) : List<LottoNumber> by originNumbers {
     val numbers: List<LottoNumber> = originNumbers.takeIf { it.isNotEmpty() } ?: LottoNumber.generateLottoNumbers()
