+++ src/test/kotlin/autolotto/vo/WinningLottoTest.kt
@@ -3,6 +3,7 @@ package autolotto.vo
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
+import lotto.vo.WinningLotto
 import java.lang.IllegalArgumentException
 
 class WinningLottoTest : FunSpec({
