+++ src/test/kotlin/autolotto/winningpoint/WinningStatisticsTest.kt
@@ -1,12 +1,14 @@
 package autolotto.winningpoint
 
-import autolotto.vo.AutoLotto
-import autolotto.vo.Lotto
-import autolotto.vo.LottoNumber
-import autolotto.vo.WinningLotto
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.maps.shouldHaveSize
 import io.kotest.matchers.shouldBe
+import lotto.vo.AutoLotto
+import lotto.vo.Lotto
+import lotto.vo.LottoNumber
+import lotto.vo.WinningLotto
+import lotto.winningpoint.WinningRank
+import lotto.winningpoint.WinningStatistics
 
 class WinningStatisticsTest : FunSpec({
     test("전달된 로또와 당첨번호를 비교하여 당첨 통계를 계산한다") {
